backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.4
django-celery-beat==2.8.1
django-elasticsearch-dsl==8.2
django-timezone-field==7.1
djangorestframework==3.16.0
elastic-transport==8.17.1
elasticsearch==8.18.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-crontab==3.3.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
